millis_t are now 64 bits and don't wrap around every couple of weeks.  buttons can be round or rectangular

+++ Types.h
@@ -11,7 +11,7 @@ enum Direction {
   DIRECTION_UP, DIRECTION_RIGHT, DIRECTION_DOWN, DIRECTION_LEFT
 };
 
-typedef unsigned long millis_t;
+typedef long long millis_t;
 typedef const char* appid_t;
 
 #endif
